requirements.txt
Django==1.11.13
psycopg2==2.7
djangorestframework
djangorestframework-jwt